fix: moved icons to skapp-ui, moved types to common file, removed extra type files
