--- frontend/src/community/crm/types/CrmCompanyTypes.ts
@@ -1,16 +1,0 @@
-export interface CrmCompanyType {
-  id: number;
-  name: string;
-  industry: string | null;
-  website: string | null;
-  address: string | null;
-  contactNumber: string | null;
-  isDeleted: boolean;
-}
-
-export interface CrmCompaniesResponseType {
-  items: CrmCompanyType[];
-  totalItems: number;
-  currentPage: number;
-  totalPages: number;
-}
