--- frontend/src/community/crm/types/CrmTaskTypes.ts
@@ -1,31 +1,0 @@
-import { CrmCompanyType } from "./CrmCompanyTypes";
-import { CrmContactType, CrmOwnerType } from "./CrmContactTypes";
-import { CrmDealType, CrmPriorityType } from "./CrmDealTypes";
-
-export interface CrmTaskCategory {
-  id: number;
-  name: string;
-  orderIndex: number;
-}
-
-export interface CrmTaskType {
-  id: number;
-  name: string;
-  type: CrmTaskCategory;
-  priority: CrmPriorityType;
-  isCompleted: boolean;
-  dueAt: string | null;
-  notes: string | null;
-  owner: CrmOwnerType;
-  contact: CrmContactType | null;
-  company: CrmCompanyType | null;
-  deal: CrmDealType | null;
-  isDeleted: boolean;
-}
-
-export interface CrmTasksResponseType {
-  items: CrmTaskType[];
-  totalItems: number;
-  currentPage: number;
-  totalPages: number;
-}
