+++ frontend/src/community/crm/types/StoreTypes.ts
